refactor(compression): extract plan-resolution leaf (planResolution.ts) under size cap (Phase 3)
